requirements.txt
Flask
flask-socketio
flask-cors
pandas
numpy
torch
transformers
peft
scikit-learn
python-docx
PyPDF2
chardet
openpyxl
nltk
scapy

# Phishing Detection Dependencies
google-auth-oauthlib
google-api-python-client
google-auth-httplib2
google-auth
beautifulsoup4
bleach
yara-python
requests
oauthlib
python-dateutil
pytz
tqdm
Pillow
langdetect
tldextract
tensorflow

